requirements.txt
anthropic>=0.49.0
gspread>=6.0.0
google-auth>=2.0.0
python-dotenv>=1.0.0
python-telegram-bot[webhooks]>=20.0
pytest>=8.0.0
playwright>=1.40.0
selectolax>=0.3.17
pyyaml>=6.0
youtube-transcript-api>=0.6.0
requests>=2.31.0
httpx>=0.27.0
